Update Chassis Elasticsearch module reference
This extension has been renamed as part of work to support
the newer Ubuntu LTS release. We need to update the name
we use to clone the project, or provisioning fails.

The other extensions on this list have yet to be updated
to use underscores instead of dashes, so this is the only
change we can make at this time.